:zap: update master data

+++ app/Http/Controllers/Api/MotorcycleController.php
@@ -9,6 +9,7 @@ use App\Helpers\ApiResponse;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CreateMotorcycleRequest;
 use App\Http\Requests\UpdateMotorcycleRequest;
+use App\Http\Resources\MotorcycleResource;
 use App\Models\DealerMotorCycle;
 use App\Models\Motorcycle;
 use Illuminate\Contracts\Database\Eloquent\Builder;
@@ -25,7 +26,7 @@ class MotorcycleController extends Controller
                 'filter.keyword' => 'string|nullable',
                 'page' => 'integer|nullable',
                 'page_size' => 'integer|between:1,100|nullable',
-                'filter.area_id' => 'integer|required',
+                'filter.area_uuid' => 'string|required',
             ]);
         } catch (ValidationException $e) {
             return ApiResponse::error($e->validator->errors(), 422, 'Validation failed');
@@ -35,9 +36,7 @@ class MotorcycleController extends Controller
         $page = $request->input('page', 1);
         $pageSize = $request->input('page_size', 10);
 
-        $query = Motorcycle::query();
-
-        $query->join('dealer_motorcycles', 'motorcycles.id', '=', 'dealer_motorcycles.motorcycle_id');
+        $query = Motorcycle::query()->with('dealers');
 
         $query->where(function (Builder $builder) use ($filter) {
             if (isset($filter['keyword'])) {
@@ -47,8 +46,8 @@ class MotorcycleController extends Controller
                         ->orWhere('color', 'like', '%' . $keyword . '%');
                 });
             }
-            if (isset($filter['area_id'])) {
-                $builder->where('area_id', $filter['area_id']);
+            if (isset($filter['area_uuid'])) {
+                $builder->where('area_uuid', $filter['area_uuid']);
             }
 
             if (isset($filter['otr'])) {
@@ -56,57 +55,42 @@ class MotorcycleController extends Controller
             }
         });
 
-        $paginator = $query->paginate($pageSize, ['*'], 'page', $page);
+        $motorcycles = MotorcycleResource::collection($query->paginate($pageSize, ['*'], 'page', $page));
 
-        $formattedData = collect($paginator->items())->map(function ($item) {
-            return [
-                'id' => $item->id,
-                'type' => $item->type,
-                'color' => $item->color,
-                'otr' => $item->otr,
-                'user_id' => $item->user_id,
-                'area_id' => $item->area_id,
-                'created_at' => $item->created_at,
-                'updated_at' => $item->updated_at,
-                'dealers' => $item->dealers->map(function ($dealer) {
-                    return [
-                        'motorcycle_id' => $dealer->pivot->motorcycle_id,
-                        'dealer_code' => $dealer->pivot->dealer_code,
-                        'dealer' => [
-                            'dealer_name' => $dealer->dealer_name,
-                            // Include other dealer information as needed
-                        ],
-                    ];
-                }),
-            ];
-        });
+        $pageInfo = $this->getPaginationDetails($motorcycles);
 
-        $pageInfo = $this->getPaginationDetails($formattedData);
-        return ApiResponse::success($formattedData->toArray(), 200, 'success', $pageInfo);
+        return ApiResponse::success($motorcycles->items(), 200, 'success', $pageInfo);
     }
 
 
     public function show($id)
     {
-        $motorcycle = Motorcycle::find($id);
-        return ApiResponse::success($motorcycle, 200, 'success');
+        $motorcycle = Motorcycle::with('dealers')->find($id);
+        if (!$motorcycle) {
+            return ApiResponse::error(null, 404, 'Motorcycle not found');
+        }
+        $resource =  new MotorcycleResource($motorcycle);
+        return ApiResponse::success($resource, 200, 'success');
     }
     public function update(UpdateMotorcycleRequest $request, $id)
     {
         try {
             $motorcycle = Motorcycle::find($id);
+            if (!$motorcycle) {
+                return ApiResponse::error(null, 404, 'Motorcycle not found');
+            }
             $motorcycle->update([
                 'type' => $request->type,
                 'color' => $request->color,
                 'otr' => $request->otr,
-                'user_id' => $request->user_id,
-                'area_id' => $request->area_id,
+                'user_uuid' => $request->user_uuid,
+                'area_uuid' => $request->area_uuid,
             ]);
             $motorcycle->dealers()->detach();
             foreach ($request->dealer_code as $dealer_code) {
                 DB::table('dealer_motorcycles')->insert([
                     'dealer_code' => $dealer_code,
-                    'motorcycle_id' => $motorcycle->id,
+                    'motorcycle_uuid' => $motorcycle->motorcycle_uuid,
                 ]);
             }
             return ApiResponse::success($motorcycle, 201, 'Motorcycle Updated');
@@ -122,14 +106,14 @@ class MotorcycleController extends Controller
             'type' => $request->type,
             'color' => $request->color,
             'otr' => $request->otr,
-            'user_id' => $request->user_id,
-            'area_id' => $request->area_id,
+            'user_uuid' => $request->user_uuid,
+            'area_uuid' => $request->area_uuid,
         ]);
 
         foreach ($request->dealer_code as $dealer_code) {
             DB::table('dealer_motorcycles')->insert([
                 'dealer_code' => $dealer_code,
-                'motorcycle_id' => $motorcycle->id,
+                'motorcycle_uuid' => $motorcycle->motorcycle_uuid,
             ]);
         }
         return ApiResponse::success($motorcycle, 201, 'success');
